requirements.txt
# =============================================================================
# requirements.txt — Affective Manifold Research Project
# Install: pip install -r requirements.txt
#
# NOTE: Install CLIP separately:
#   pip install git+https://github.com/openai/CLIP.git
#
# NOTE: GraphRicciCurvature:
#   pip install GraphRicciCurvature
# =============================================================================

# Core scientific stack
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-learn-extra>=0.3.0      # for KMedoids

# Optimal transport
POT>=0.9.0                     # Python Optimal Transport (ot library)

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# NLP / embeddings
sentence-transformers>=2.2.0   # includes all-mpnet-base-v2

# Graph tools
networkx>=3.0
GraphRicciCurvature>=0.5.3     # Ollivier-Ricci computation

# Statistics
statsmodels>=0.14.0
pingouin>=0.5.3                # mediation analysis
scipy>=1.10.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3              # UMAP visualisations

# Config and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Anthropic API (OPTIONAL — only if query_parser = "anthropic" in config.yaml)
anthropic>=0.20.0             # optional — only needed for Claude API query parser

# Clustering
scikit-learn-extra>=0.3.0

# BH-FDR correction (via statsmodels)
# statsmodels already listed above
